requirements.txt
astroid==2.0.4
Django==2.1.2
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.5
selenium==3.14.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
